TK-00000
Fixed bug where BufferedReadCapacity16Task was writing block size as a Long instead of an Int

git-svn-id: https://jscsi.svn.sourceforge.net/svnroot/jscsi/trunk@384 6391b2ca-d626-0410-90c9-8f6d1efc1dcd

+++ projects/scsi/src/org/jscsi/scsi/tasks/buffered/BufferedReadCapacity16Task.java
@@ -38,7 +38,7 @@ public class BufferedReadCapacity16Task extends BufferedTask
       try
       {
          out.writeLong(this.getFileCapacity());
-         out.writeLong((long)blockLength);
+         out.writeInt(blockLength);
          out.writeByte(0);    // RTO_EN and PROT_EN set to false; do not support protection info
          // the remaining bytes are reserved
       }
